requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.1
email-validator==2.2.0 